added schema and started on seeding document
